put thread init mostly back to how it was, changing setbits implementation

+++ os_setbits.h
@@ -15,13 +15,13 @@ int os_setbits_init(os_setbits_t *mod);
  * @brief Signals a specific bit
  * @param os_setbits_t *pointer to module
  */
-int os_setbits_signal(os_setbits_t *mod, uint8_t bit);
+int os_setbits_signal(os_setbits_t *mod, int bit);
 
 /**
  * @brief Clears a specific bit
  * @param os_setbits_t *pointer to module
  */
-int os_clearbits(os_setbits_t *mod, uint8_t bit);
+int os_clearbits(os_setbits_t *mod, int bit);
 
 /**
  * @brief Wait for a specific bit to be set
@@ -33,6 +33,6 @@ int os_waitbits(os_setbits_t *mod, uint8_t bit, uint32_t timeout_ms);
  * @brief Wait indefinitely for a specific bit to be set
  * @param os_setbits_t *pointer to module
  */
-int os_waitbits_indefinite(os_setbits_t *mod, uint8_t bit);
+int os_waitbits_indefinite(os_setbits_t *mod, int bit);
 
 #endif
